Replace an old segtree by our new one
