+++ crates/algolib/segtree/src/tests/tester.rs
