--- crates/algolib/segtree2/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "segtree2"
-version = "0.1.0"
-authors = ["ngtkana <ngtkana@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-type_traits = { path = "../../utils/type_traits" }
-
-[dev-dependencies]
-fp = { path = "../../utils/fp" }
-query_test = { path = "../../utils/query_test" }
-rand = "0.7.3"
